requirements.txt
pydantic==1.7.2
torchtext==0.8.1
ijson==3.1.3
stanza==1.1.1
h5py==3.1.0
orjson==3.4.6
